Make EventLoop.schedule_* functions always return an unschedule handle.
schedule_in() and schedule_periodically() will now always return handles
that can be passed to unschedule(), even when not ran from the event
loop thread. In addition, calling unschedule() with the handle for
periodical callbacks will now also work after the first call of that
callback. The opaque IDs are currently implemented as a simple
monotonically increasing number.

This fixes #19.